+++ js/data.js
@@ -1,0 +1,47 @@
+const OFFER_TITLE = 'Жилье вашей мечты';
+
+const OFFER_DESCRIPTION = 'Уютные апартаменты в центре города по привлекательно цене';
+
+const OFFER_TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+
+const OFFER_CHECKIN = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const OFFER_CHECKOUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const OFFER_FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const OFFER_PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+export {
+  OFFER_TITLE,
+  OFFER_DESCRIPTION,
+  OFFER_TYPE,
+  OFFER_CHECKIN,
+  OFFER_CHECKOUT,
+  OFFER_FEATURES,
+  OFFER_PHOTOS
+};
